Moves prior exposure processing into its own module
